+++ packages/ui/src/Button/index.ts
@@ -1,10 +1,3 @@
-export {
-    ButtonDefault,
-    ButtonPrimary,
-    ButtonSecondary,
-    ButtonFloating,
-    ButtonIcon
-} from "./Button";
-
+export * from "./Button";
 export * from "./IconButton/IconButton";
 export * from "./CopyButton/CopyButton";