+++ packages/api-headless-cms/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.12.1](https://github.com/webiny/webiny-js/compare/v4.12.1-beta.0...v4.12.1) (2020-09-17)
+
+**Note:** Version bump only for package @webiny/api-headless-cms
+
+
+
+
+
+## [4.12.1-beta.0](https://github.com/webiny/webiny-js/compare/v4.12.0...v4.12.1-beta.0) (2020-09-17)
+
+**Note:** Version bump only for package @webiny/api-headless-cms
+
+
+
+
+
+# [4.12.0](https://github.com/webiny/webiny-js/compare/v4.12.0-beta.1...v4.12.0) (2020-09-16)
+
+**Note:** Version bump only for package @webiny/api-headless-cms
+
+
+
+
+
+# [4.12.0-beta.1](https://github.com/webiny/webiny-js/compare/v4.12.0-beta.0...v4.12.0-beta.1) (2020-09-16)
+
+**Note:** Version bump only for package @webiny/api-headless-cms
+
+
+
+
+
+# [4.12.0-beta.0](https://github.com/webiny/webiny-js/compare/v4.11.0...v4.12.0-beta.0) (2020-09-16)
+
+
+### Bug Fixes
+
+* add `createSchema` prop to `cms-model-field-to-graphql` plugin's `read` property ([67a3b36](https://github.com/webiny/webiny-js/commit/67a3b36c61faa6c1f4a88f002037d7c33529bcd0))
+* ensure multiple `models` context plugins don't clash ([044ee77](https://github.com/webiny/webiny-js/commit/044ee77bf396bda33624b74a41b4c560ff843693))
+
+
+
+
+
+# [4.11.0](https://github.com/webiny/webiny-js/compare/v4.11.0-beta.1...v4.11.0) (2020-09-09)
+
+
+### Bug Fixes
+
+* **api-headless-cms:** add equality check for id `idValidation` ([5e7ee98](https://github.com/webiny/webiny-js/commit/5e7ee98879e52a799d86a7a8ddd6e0b8baef5f84))
+* **api-headless-cms:** trim `fieldId` before save ([173db34](https://github.com/webiny/webiny-js/commit/173db349bfb2fb1f08fc51345cf5bd01732c9851))
+
+
+
+
+
 # [4.11.0-beta.1](https://github.com/webiny/webiny-js/compare/v4.11.0-beta.0...v4.11.0-beta.1) (2020-09-09)
 
 **Note:** Version bump only for package @webiny/api-headless-cms